Moving tests/Transport to tests/AjaxTransport and tests/NodeTransport.
